pyproject.toml
[project]
name = "weather-agent-mcp"
version = "0.1.0"
description = "Weather agent with MCP server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn==0.32.1",
    "openai==1.57.2",
    "pydantic==2.11.5",
    "fastmcp==0.3.0",
    "python-dotenv==1.0.1",
]

[dependency-groups]
dev = [
    "ruff",
    "pyright",
    "black",
]
